leetcode(4): simplify existing code
Simple for now. I'll analyse the time complexity after implementing the
initial solution, and determine if it's better to concatenate the two
lists, or to try to do what I was previously attempting: calculate the
median index by halving the sum of the lengths of the two arrays, and
then determining if that median index falls in the left or right array,
and returning it.