leetcode(409): correct palindrome generation
Forcing remaining_count to be even is pointless, it's already half of an
even number, and anything doubled is even. Also, subtracting one
conditionally means a conditional addition has to be made later.

Additionally, unused_letters was being calculated incorrectly: set
difference isn't an associative operation, and so unused_letters was
previously expressing the set of all letters used by the palindrome that
weren't in LETTERS, which will always be an empty set.

Lastly, it's possible for the palindrome to use all letters, so if there
are no letters left, best bet is to add one of the already used letters,
or to decrement the correct_answer.