chore(elm): consider discontinuing Elm
The state of Elm seems to be a bit uncertain:

https://discourse.elm-lang.org/t/request-elm-0-19-2-any-update-to-help-adoption-to-prove-that-elm-is-not-dead/8843/15

At the time of writing, the statements above the linked one are still
correct: the compiler hasn't seen major (or minor) changes in ~5 years.
The distribution of Elm was updated, and commits have been made to the
main compiler repo, but these are relatively minor when considered
through the lens that modern software development puts upon open source
project velocity metrics.

Gren was forked from Elm a while ago.

The discussion here is useful, to be sure:

https://www.reddit.com/r/elm/comments/y6hgwx/is_elm_still_being_developed/

I personally don't have many expectations from Elm, and it would likely
work really quite well for the size of projects I plan to do.

I think for now I'll look at other languages, to see if something else
catches my fancy.

Current tabs for Elm:

https://elmcraft.org/learn/
https://elmprogramming.com/building-a-simple-page-in-elm.html
https://elm-lang.org/docs/syntax#conditionals
https://guide.elm-lang.org/architecture/forms
https://package.elm-lang.org/packages/elm/core/latest/Basics
https://package.elm-lang.org/packages/elm/html/1.0.0/Html-Attributes
https://package.elm-lang.org/packages/elm/html/1.0.0/Html
https://www.youtube.com/watch?v=6TDKHGtAxeg&t=85s
https://www.youtube.com/watch?v=IcgmSRJHu_8&t=74s
https://exercism.org/tracks/elm/exercises/lucians-luscious-lasagna
https://www.youtube.com/watch?v=XpDsk374LDE
https://incrementalelm.com/
https://elm.land/
https://www.reddit.com/r/elm/comments/ryykkc/is_elm_dead/
https://discourse.elm-lang.org/t/status-update-3-nov-2021/7870
https://www.reddit.com/r/elm/comments/12961nu/comment/jf5ebb6/
https://news.ycombinator.com/item?id=34746161
https://lukeplant.me.uk/blog/posts/why-im-leaving-elm/#entitlement-and-expectations
https://dev.to/kspeakman/elm-019-broke-us--khn
https://discourse.elm-lang.org/t/native-code-in-0-19/826
https://news.ycombinator.com/item?id=16510267